添加玩家昵称支持，Summary增加更新时间字段
Former-commit-id: 44f2c16f965665eaf088e96f213ae9fa98baa746 [formerly 44f2c16f965665eaf088e96f213ae9fa98baa746 [formerly 9df36a1859c944b05e5da8679c2a135a59f01c58]]
Former-commit-id: 532d93268cd3bb504f3f2d44b5a959d787df3c60
Former-commit-id: c67ff6179e0cfbf3d053b9aa40569a99329ba448

+++ PhigrosLibrary/src/main/java/given/phigros/GameProgress.java
@@ -8,9 +8,9 @@ public class GameProgress implements SaveModule {
     public boolean alreadyShowCollectionTip;
     public boolean alreadyShowAutoUnlockINTip;
     public String completed;
-    public int songUpdateInfo;
+    public byte songUpdateInfo;
     public short challengeModeRank;
-    public int[] money = new int[5];
+    public short[] money = new short[5];
     public byte unlockFlagOfSpasmodic;
     public byte unlockFlagOfIgallta;
     public byte unlockFlagOfRrharil;
